requirements.txt
streamlit>=1.31.0
plotly>=5.18.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
openpyxl>=3.1.0
fpdf2>=2.7.0
